move host IRs related files to a separate folder
