+++ CMakeLists.txt
@@ -124,9 +124,9 @@ list(APPEND NVFUSER_SRCS
   ${NVFUSER_SRCS_DIR}/fusion_segmenter.cpp
   ${NVFUSER_SRCS_DIR}/global_allocator.cpp
   ${NVFUSER_SRCS_DIR}/grouped_reduction.cpp
-  ${NVFUSER_SRCS_DIR}/host_ir_container.cpp
-  ${NVFUSER_SRCS_DIR}/host_ir_executor.cpp
-  ${NVFUSER_SRCS_DIR}/ir/host_ir.cpp
+  ${NVFUSER_SRCS_DIR}/host_ir/container.cpp
+  ${NVFUSER_SRCS_DIR}/host_ir/executor.cpp
+  ${NVFUSER_SRCS_DIR}/host_ir/host_ir.cpp
   ${NVFUSER_SRCS_DIR}/id_model/id_model.cpp
   ${NVFUSER_SRCS_DIR}/id_model/to_string.cpp
   ${NVFUSER_SRCS_DIR}/id_model/transform_replay.cpp
@@ -497,7 +497,6 @@ list(APPEND JIT_TEST_SRCS
   ${NVFUSER_ROOT}/tests/cpp/test_id_model.cpp
   ${NVFUSER_ROOT}/tests/cpp/test_iter_visitor.cpp
   ${NVFUSER_ROOT}/tests/cpp/test_linked_hash_map.cpp
-  ${NVFUSER_ROOT}/tests/cpp/test_host_irs.cpp
   ${NVFUSER_ROOT}/tests/cpp/test_loop_rotation.cpp
   ${NVFUSER_ROOT}/tests/cpp/test_mbarrier.cpp
   ${NVFUSER_ROOT}/tests/cpp/test_memory.cpp
@@ -602,6 +601,9 @@ if(BUILD_TEST)
 
   add_test(tutorial "${NVFUSER_ROOT}/tests/cpp/test_tutorial.cpp" "")
   list(APPEND TEST_BINARIES tutorial)
+
+  add_test(test_host_ir "${NVFUSER_ROOT}/tests/cpp/test_host_irs.cpp" "")
+  list(APPEND TEST_BINARIES test_host_ir)
 
   if(BUILD_PYTHON)
     set(PY_FRONTEND_TEST_SRCS)
