+++ csrc/host_ir/container.cpp
@@ -6,10 +6,10 @@
  */
 // clang-format on
 
-#include <host_ir_container.h>
+#include <host_ir/container.h>
 #include <ir/builder.h>
 #include <ir/cloner.h>
-#include <ir/host_ir.h>
+#include <host_ir/host_ir.h>
 #include <ir/printer.h>
 #include <ir/utils.h>
 #include <kernel_ir.h>
