+++ csrc/host_ir/executor.h
@@ -9,8 +9,8 @@
 
 #include <dispatch.h>
 #include <executor.h>
-#include <host_ir_container.h>
-#include <ir/host_ir.h>
+#include <host_ir/container.h>
+#include <host_ir/host_ir.h>
 #include <kernel_cache.h>
 
 namespace nvfuser {
