+++ csrc/host_ir/host_ir.cpp
@@ -6,10 +6,10 @@
  */
 // clang-format on
 
-#include <host_ir_container.h>
+#include <host_ir/container.h>
+#include <host_ir/host_ir.h>
 #include <ir/builder.h>
 #include <ir/cloner.h>
-#include <ir/host_ir.h>
 #include <ir/printer.h>
 #include <ir/utils.h>
 #include <kernel_ir.h>
@@ -21,7 +21,7 @@ namespace hir {
 
 HostUnit::HostUnit(IrBuilderPasskey passkey, std::unique_ptr<Fusion> fusion)
     : Expr(passkey), fusion_(std::make_unique<Fusion>(*fusion)) {
-  NVF_ERROR(passkey.ir_container_->isA<hir::HostIrContainer>());
+  NVF_ERROR(passkey.ir_container_->isA<hir::HostIrContainer>()); //NOLINT
 }
 
 HostUnit::HostUnit(const HostUnit* src, IrCloner* ir_cloner)
@@ -52,7 +52,7 @@ PostOnStream::PostOnStream(
     std::vector<Val*> inputs,
     std::vector<Val*> outputs)
     : Expr(passkey, std::move(inputs), std::move(outputs), {hu}) {
-  NVF_ERROR(passkey.ir_container_->isA<hir::HostIrContainer>());
+  NVF_ERROR(passkey.ir_container_->isA<hir::HostIrContainer>()); //NOLINT
   NVF_ERROR(this->inputs().size() == hu->fusion_to_execute()->inputs().size());
   NVF_ERROR(
       this->outputs().size() == hu->fusion_to_execute()->outputs().size());
