+++ csrc/ir/all_nodes.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <ir/base_nodes.h>
-#include <ir/host_ir.h>
 #include <ir/interface_nodes.h>
 #include <ir/internal_nodes.h>
 #include <kernel_ir.h>
+#include <host_ir/host_ir.h>
